backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
pytesseract
pillow
numpy
